old-tests/submission/Microsoft/video/video_018.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'playbackRate' returns 1.0 if it has not been set.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-defaultplaybackrate" />
  <meta name="assert" content="'playbackRate' returns 1.0 if it has not been set." />

</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <div id=log></div>
    <script type="text/javascript">
        test(function() { assert_equals(document.createElement("video").playbackRate, 1.0)}, "'playbackRate' returns 1.0 if it has not been set.");
    </script>
</div>
</body>
</html>
